Update permanent redirections
